Test converted location ordering in RecentLocationsTest
